Move config.py to project's root directory
